e2e_test: add temporary workaround to copy Mist testing binaries to CI host
The Mist tester binaries need to run outside of the containers or inside another
container. This is a temporary workaround to run outside the containers directly
on the CI host. This is obviously a bit fragile and a follow up change will be used
to make this more robust by using a third container that contains all relevant test
binaries