e2e_test: set stream-interruption test to fail correctly
In #86, we added stream interruption test i.e. verify playback still recovers
after ingest stream is killed and restarted. However, we set it to forcefully
pass even with failures - this was temporary until we root-caused the issue.
With #65 now resolved, this stream interruption test can be correctly set to
fail if playback is not able to recover correctly during an ingest stream
interruption.